backend_old/requirements.txt
flask==2.0.2
flask_cors==3.0.10
pigpio-dht==0.3.6
pigpio==1.78
pyyaml==6.0